Update docs for material, split up light classes.
