+++ docs/layers/grid-layer.md
@@ -189,10 +189,10 @@ When set to true, aggregation is performed on GPU, provided other conditions are
 
 ##### `material` (Object, optional)
 
-* Default: `new PhongMaterial()`
+* Default: `true`
 
 This is an object that contains material props for [lighting effect](/docs/effects/lighting-effect.md) applied on extruded polygons.
-Check [PhongMaterial](https://github.com/uber/luma.gl/tree/7.0-release/docs/api-reference/core/materials/phong-material.md) for more details.
+Check [the lighting guide](/docs/developer-guide/using-lighting.md) for more details.
 
 
 ### Data Accessors
