+++ docs/layers/point-cloud-layer.md
@@ -93,10 +93,10 @@ Global radius of all points, in units specified by `sizeUnits` (default pixels).
 
 ##### `material` (Object, optional)
 
-* Default: `new PhongMaterial()`
+* Default: `true`
 
-This is an object that contains material props for [lighting effect](/docs/effects/lighting-effect.md).
-Check [PhongMaterial](https://github.com/uber/luma.gl/tree/7.0-release/docs/api-reference/core/materials/phong-material.md) for more details.
+This is an object that contains material props for [lighting effect](/docs/effects/lighting-effect.md) applied on extruded polygons.
+Check [the lighting guide](/docs/developer-guide/using-lighting.md) for more details.
 
 ### Data Accessors
