+++ docs/layers/polygon-layer.md
@@ -186,10 +186,10 @@ Only works if `getLineDashArray` is specified.
 
 ##### `material` (Object, optional)
 
-* Default: `new PhongMaterial()`
+* Default: `true`
 
 This is an object that contains material props for [lighting effect](/docs/effects/lighting-effect.md) applied on extruded polygons.
-Check [PhongMaterial](https://github.com/uber/luma.gl/tree/7.0-release/docs/api-reference/core/materials/phong-material.md) for more details.
+Check [the lighting guide](/docs/developer-guide/using-lighting.md) for more details.
 
 ### Data Accessors
 
@@ -267,7 +267,7 @@ The width of the outline of the polygon, in units specified by `lineWidthUnits` 
 
 * Default: `1000`
 
-The elevation to extrude each polygon with. 
+The elevation to extrude each polygon with.
 If a cartographic projection mode is used, height will be interpreted as meters,
 otherwise will be in unit coordinates.
 Only applies if `extruded: true`.
