+++ docs/layers/solid-polygon-layer.md
@@ -99,10 +99,10 @@ all elevation without updating the data.
 
 ##### `material` (Object, optional)
 
-* Default: `new PhongMaterial()`
+* Default: `true`
 
 This is an object that contains material props for [lighting effect](/docs/effects/lighting-effect.md) applied on extruded polygons.
-Check [PhongMaterial](https://github.com/uber/luma.gl/tree/7.0-release/docs/api-reference/core/materials/phong-material.md) for more details.
+Check [the lighting guide](/docs/developer-guide/using-lighting.md) for more details.
 
 ### Data Accessors
 
@@ -145,7 +145,7 @@ Only applies if `extruded: true`.
 
 * Default: `1000`
 
-The elevation to extrude each polygon with. 
+The elevation to extrude each polygon with.
 If a cartographic projection mode is used, height will be interpreted as meters,
 otherwise will be in unit coordinates.
 Only applies if `extruded: true`.
