+++ modules/core/src/effects/lighting/lighting-effect.js
@@ -1,5 +1,6 @@
 import {Texture2D, ProgramManager} from '@luma.gl/core';
-import {AmbientLight, DirectionalLight} from './lights';
+import {AmbientLight} from './ambient-light';
+import {DirectionalLight} from './directional-light';
 import Effect from '../../lib/effect';
 import {Matrix4, Vector3} from 'math.gl';
 import ShadowPass from '../../passes/shadow-pass';
